refactor(eslint-plugin-react-hooks): move rules to `rules` folder
Since the compiler plugin is going to be merged into the hooks plugin,, and ultimately decomposed into several more rules, it would be good to start creating a more traditional folder structure for the plugin.  This change just moves the rules into a `rules` folder.
