+++ packages/eslint-plugin-react-hooks/src/rules/RulesOfHooks.ts
